exeption handling and response entities were added to project

+++ src/main/java/com/everest/employee_portal/services/EmployeeService.java
@@ -1,13 +1,13 @@
 package com.everest.employee_portal.services;
 
 import com.everest.employee_portal.entities.Employee;
+import com.everest.employee_portal.exceptions.EmployeeNotFoundException;
 import com.everest.employee_portal.repositories.AddressRepository;
 import com.everest.employee_portal.repositories.EmployeeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,18 +24,24 @@ public class EmployeeService {
     }
 
     public Optional<Employee> getByID(Long id) {
-        return employeeRepository.findById(id);
+        Optional<Employee> employee = employeeRepository.findById(id);
+        if (employee.isEmpty()) {
+            throw new EmployeeNotFoundException();
+        }
+        return employee;
     }
 
-    public Long create(Employee employee) {
+    public Employee create(Employee employee) {
         addressRepository.save(employee.getPermanentAddress());
         addressRepository.save(employee.getPresentAddress());
-        Employee createdEmployee = employeeRepository.save(employee);
-        return createdEmployee.getId();
+        return employeeRepository.save(employee);
     }
 
-    public Long update(Employee employee,Long employeeId) {
+    public Employee update(Employee employee, Long employeeId) {
         Employee employee1 = employeeRepository.getById(employeeId);
+        if (employee1.id == null) {
+            throw new EmployeeNotFoundException();
+        }
         Long permanentAddressId = employee1.getPermanentAddress().getId();
         Long presentAddressId = employee1.getPresentAddress().getId();
         employee.getPermanentAddress().setId(permanentAddressId);
@@ -43,19 +49,21 @@ public class EmployeeService {
         addressRepository.save(employee.getPermanentAddress());
         addressRepository.save(employee.getPresentAddress());
         employeeRepository.save(employee);
-        return employee.getId();
+        return employee;
     }
 
-    public String delete(Long id) {
+    public Employee delete(Long id) {
         Employee employee = employeeRepository.getById(id);
+        if(employee.id==null){
+            throw new EmployeeNotFoundException();
+        }
         addressRepository.delete(employee.getPermanentAddress());
         addressRepository.delete(employee.getPresentAddress());
         employeeRepository.delete(employee);
-        return "Employee with " + id + "deleted successfully";
+        return employee;
     }
 
-    public List<Employee> searchEmployeeWith(String firstName, String lastName, String dateOfJoin) {
-        System.out.println(dateOfJoin + " in service class");
-       return employeeRepository.findByFirstNameAndLastName(firstName,lastName);
+    public List<Employee> searchEmployeeWith(String firstName, String lastName) {
+        return employeeRepository.findByFirstNameAndLastName(firstName, lastName);
     }
 }
